requirements.txt
streamlit==1.32.0
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
scikit-learn==1.3.2
scikit-image==0.22.0
joblib==1.3.2
tqdm==4.66.4
matplotlib==3.8.3
